Replace the unreachable fused instruction with a new one.
1. Add new condition that destination register of instruction is equal to
its source register, because the previous condition does not update the
value of destination register of the first instruction.
2. Based on experiment and document of ria-jit, the fused instruciton auipc
+ add is unreachable, so we replace it with new fused instrucion lui +
addi.

+++ src/decode.h
@@ -253,7 +253,7 @@ typedef struct rv_insn {
     uint8_t shamt;
 #endif
     /* fuse operation */
-    int16_t imm2;
+    int32_t imm2;
     opcode_fuse_t *fuse;
 
     /* instruction length */
